# api.yaml
swagger: "2.0.0"
info:
  title: Academy APP
  description: Api Documentation for Academia Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:
...
  /v1/subject/createSubject:
    post:
      tags:
        - name: Subject
      summary: Create Subject
      description: Api to Create Subject
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: name
          in: formData
          type: string
          required: true
        - name: image
          in: formData
          type: file
        - name: status
          in: formData
          type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/Subject/editSubject/{subjectID}:
    patch:
      tags:
        - name: Subject
      summary: Edit the Subject
      description: Api that will edit the Subject
      consumes: application/json
      produces: application/json
      parameters:
        - name: subjectID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
        - name: status
          in: formData
          type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subject/deleteSubjectById/{subjectId}:
    delete:
      tags:
        - name: Subject
      summary: Delete Subject by its id
      description: APi to delete subject by id
      parameters:
        - name: subjectId
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error

  #### ------------- Board Route ------------- ####
...
  /v1/subBoard/createSubBoard/{boardID}:
    post:
      tags:
        - name: SubBoard
      summary: Create subBoard
      description: Api to Create subBoard
      consumes: application/json
      produces: application/json
      parameters:
        - name: name
          in: formData
          type: string
          required: true
        - name: boardID
          in: path
          type: string
          required: true
        - name: status
          in: formData
          type: string
          enum: [active, inactive]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subBoard/editSubBoard/{subBoardID}:
    patch:
      tags:
        - name: SubBoard
      summary: Edit the subBoard
      description: Api that will edit the subBoard
      consumes: application/json
      produces: application/json
      parameters:
        - name: subBoardID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
        - name: status
          in: formData
          type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subBoard/deleteSubBoardById/{subBoardID}:
    delete:
      tags:
        - name: SubBoard
      summary: Delete SubBoard by its id
      description: APi to delete SubBoard by id
      parameters:
        - name: subBoardID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  # #### ------------- Qualification Route ------------- ####
  #   /v1/qualification/filterQualification:
  #     get:
  #       tags:
  #         - name: Qualification
  #       summary: Get All Qualification By filter
  #       description: APi to get all the Qualification by filter
  #       parameters:
  #         - name: searchString
  #           in: query
  #           type: string
  #         - name: status
  #           in: query
  #           type: string
  #           enum: [active, inactive]
  #         - name: startDate
  #           in: query
  #           type: string
  #         - name: endDate
  #           in: query
  #           type: string
  #         - name: page
  #           in: query
  #           type: string
  #         - name: limit
  #           in: query
  #           type: string
  #         - name: sortByDate
  #           in: query
  #           type: string
  #           enum: [asc, dsc]
  #       responses:
  #         200:
  #           description: OK
  #         400:
  #           description: Invalid Credentials
  #         404:
  #           description: Not Found
  #         401:
  #           description: Invalid Credentials
  #         500:
  #           description: Internal Server Error
  #   /v1/qualification/getQualificationById/{qualificationID}:
  #     get:
  #       tags:
  #         - name: Qualification
  #       summary: Get All qualifications By qualificationID
  #       description: APi to get all the Qualifications
  #       parameters:
  #         - name: qualificationID
  #           in: path
  #           type: string
  #           required: true
  #       responses:
  #         200:
  #           description: OK
  #         400:
  #           description: Invalid Credentials
  #         404:
  #           description: Not Found
  #         401:
  #           description: Invalid Credentials
  #         500:
  #           description: Internal Server Error

  #   /v1/qualification/createQualification:
  #     post:
  #       tags:
  #         - name: Qualification
  #       summary: Create Qualification
  #       description: Api to Create Qualification
  #       consumes: application/json
  #       produces: application/json
  #       parameters:
  #         - name: name
  #           in: formData
  #           type: string
  #           required: true
  #         - name: status
  #           in: formData
  #           type: string
  #           enum: [active, inactive]
  #       responses:
  #         200:
  #           description: OK
  #         400:
  #           description: Invalid Credentials
  #         404:
  #           description: Not Found
  #         401:
  #           description: Invalid Credentials
  #         500:
  #           description: Internal Server Error
  #   /v1/Qualification/editQualification/{qualificationID}:
  #     patch:
  #       tags:
  #         - name: Qualification
  #       summary: Edit the Qualification
  #       description: Api that will edit the Qualification
  #       consumes: application/json
  #       produces: application/json
  #       parameters:
  #         - name: qualificationID
  #           in: path
  #           type: string
  #           required: true
  #         - name: name
  #           in: formData
  #           type: string
  #         - name: status
  #           in: formData
  #           type: string
  #       responses:
  #         200:
  #           description: OK
  #         400:
  #           description: Invalid Credentials
  #         404:
  #           description: Not Found
  #         401:
  #           description: Invalid Credentials
  #         500:
  #           description: Internal Server Error
  #   /v1/qualification/deleteQualificationById/{qualificationId}:
  #     delete:
  #       tags:
  #         - name: Qualification
  #       summary: Delete Qualification by its id
  #       description: APi to delete qualification by id
  #       parameters:
  #         - name: qualificationId
  #           in: path
  #           type: string
  #           required: true
  #       responses:
  #         200:
  #           description: OK
  #         400:
  #           description: Invalid Credentials
  #         404:
  #           description: Not Found
  #         401:
  #           description: Invalid Credentials
  #         500:
  #           description: Internal Server Error

  #### ------------- Classes Route ------------- ####
...
  /v1/subTopic/filterSubTopic:
    get:
      tags:
        - name: SubTopic
      summary: Get All SubTopics By filter
      description: APi to get all the SubTopics by filter
      parameters:
        - name: searchString
          in: query
          type: string
        - name: topicID
          in: query
          type: string
        - name: status
          in: query
          type: string
          enum: [active, inactive]
        - name: startDate
          in: query
          type: string
        - name: endDate
          in: query
          type: string
        - name: page
          in: query
          type: string
        - name: limit
          in: query
          type: string
        - name: sortByDate
          in: query
          type: string
          enum: [asc, dsc]

      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subTopic/getSubTopicById/{subTopicID}:
    get:
      tags:
        - name: SubTopic
      summary: Get All SubTopic By ID
      description: API to get a SubTopic by ID
      parameters:
        - name: subTopicID
          in: path
          type: string
          required: true

      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subTopic/createSubTopic/{topicID}:
    post:
      tags:
        - name: SubTopic
      summary: Create SubTopic
      description: Api to Create SubTopic
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: topicID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
          required: true
        - name: description
          in: formData
          type: string
          required: true
        - name: file
          in: formData
          type: file
          required: false
        - name: status
          in: formData
          type: string
          enum: [active, inactive]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subTopic/editSubTopic/{subTopicID}:
    patch:
      tags:
        - name: SubTopic
      summary: Edit SubTopic
      description: Api to Edit SubTopic
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: subTopicID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
        - name: description
          in: formData
          type: string
        - name: file
          in: formData
          type: file
        - name: status
          in: formData
          type: string
          enum: [active, inactive]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/subTopic/deleteSubTopicById/{subTopicID}:
    delete:
      tags:
        - name: SubTopic
      summary: Delete SubTopic
      description: Delete SubTopic by its id
      parameters:
        - name: subTopicID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  # Exam Routes
...